migrate asyncopenai
