+++ toponymy/tests/test_async_llm_wrappers.py
@@ -99,12 +99,12 @@ def test_supported_async_namers_do_not_warn_on_callback(namer_cls, kwargs):
 
     with warnings.catch_warnings(record=True) as record:
         warnings.simplefilter("always")
-        namer_cls(callback=callback, **kwargs)
+        namer = namer_cls(callback=callback, **kwargs)
 
     debug_warnings = [w for w in record if "debug callback" in str(w.message)]
 
     assert len(debug_warnings) == 0
-    assert namer_cls._supports_debug_callback is True
+    assert namer._supports_debug_callback is True
 
 
 @pytest.mark.parametrize("namer_cls, kwargs", UNSUPPORTED_ASYNC_DEBUG_CALLBACK_NAMERS)
@@ -112,8 +112,8 @@ def test_unsupported_async_namers_warn_on_callback(namer_cls, kwargs):
     callback = lambda payload: None
 
     with pytest.warns(UserWarning, match="debug callback") as record:
-        namer_cls(callback=callback, **kwargs)
-    assert namer_cls._supports_debug_callback is False
+        namer = namer_cls(callback=callback, **kwargs)
+    assert namer._supports_debug_callback is False
 
 
 @pytest.mark.parametrize(
@@ -562,16 +562,6 @@ async def test_batch_anthropic_wait_for_completion(batch_anthropic_wrapper):
 
 
 # AsyncOpenAI Tests
-@pytest_asyncio.fixture
-async def async_openai_wrapper():
-    mock_client = AsyncMock()
-    mock_client.close = AsyncMock()
-    with patch("openai.AsyncOpenAI", return_value=mock_client):
-        wrapper = AsyncOpenAINamer(api_key="dummy")
-        try:
-            yield wrapper
-        finally:
-            await wrapper.close()
 
 
 @pytest.mark.external
@@ -611,201 +601,6 @@ async def test_openai_connectivity_async_system_canary():
         f"Async system canary failed for OpenAI:\n"
         f"  Error: {result['error_type']}: {result['error_message']}"
     )
-
-
-@pytest.mark.asyncio
-async def test_async_openai_generate_topic_names_success(
-    async_openai_wrapper, mock_data
-):
-    response = MockAsyncResponse.create_chat_response(mock_data["valid_topic_name"])
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        return_value=response
-    )
-
-    result = await async_openai_wrapper.generate_topic_names(["test prompt"])
-    assert len(result) == 1
-    validate_topic_name(result[0])
-
-
-@pytest.mark.asyncio
-async def test_async_openai_generate_topic_names_system_prompt(
-    async_openai_wrapper, mock_data
-):
-    response = MockAsyncResponse.create_chat_response(mock_data["valid_topic_name"])
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        return_value=response
-    )
-
-    result = await async_openai_wrapper.generate_topic_names(
-        [{"system": "system prompt", "user": "test prompt"}]
-    )
-    assert len(result) == 1
-    validate_topic_name(result[0])
-
-
-@pytest.mark.asyncio
-async def test_async_openai_generate_topic_cluster_names_success(
-    async_openai_wrapper, mock_data
-):
-    response = MockAsyncResponse.create_chat_response(mock_data["valid_cluster_names"])
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        return_value=response
-    )
-
-    result = await async_openai_wrapper.generate_topic_cluster_names(
-        ["test prompt"], [mock_data["old_names"]]
-    )
-    assert len(result) == 1
-    validate_cluster_names(result[0])
-
-
-@pytest.mark.asyncio
-async def test_async_openai_generate_topic_cluster_names_system_prompt(
-    async_openai_wrapper, mock_data
-):
-    response = MockAsyncResponse.create_chat_response(mock_data["valid_cluster_names"])
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        return_value=response
-    )
-
-    result = await async_openai_wrapper.generate_topic_cluster_names(
-        [{"system": "system prompt", "user": "test prompt"}], [mock_data["old_names"]]
-    )
-    assert len(result) == 1
-    validate_cluster_names(result[0])
-
-
-@pytest.mark.asyncio
-async def test_async_openai_generate_topic_cluster_names_malformed_mapping(
-    async_openai_wrapper, mock_data
-):
-    response = MockAsyncResponse.create_chat_response(mock_data["malformed_mapping"])
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        return_value=response
-    )
-
-    result = await async_openai_wrapper.generate_topic_cluster_names(
-        ["test prompt"], [mock_data["old_names"]]
-    )
-    assert len(result) == 1
-    validate_cluster_names(result[0])
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("error_class", OPENAI_FAIL_FAST)
-async def test_async_openai_generate_topic_names_fail_fast_raises(
-    async_openai_wrapper, error_class
-):
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        side_effect=make_openai_error(error_class)
-    )
-
-    with pytest.raises(FailFastLLMError):
-        await async_openai_wrapper.generate_topic_names(["test prompt"])
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("error_class", OPENAI_RETRYABLE)
-@pytest.mark.filterwarnings("ignore:Failed to generate")
-async def test_async_openai_generate_topic_names_retryable_returns_empty(
-    async_openai_wrapper,
-    error_class,
-):
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        side_effect=[make_openai_error(error_class) for _ in range(3)]
-    )
-
-    result = await async_openai_wrapper.generate_topic_names(["test prompt"])
-
-    assert result == [""]
-    assert async_openai_wrapper.client.chat.completions.create.call_count == 3
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("error_class", OPENAI_RETRYABLE)
-@pytest.mark.filterwarnings("ignore:Failed to generate")
-async def test_async_openai_generate_topic_cluster_names_retryable_returns_old_names(
-    async_openai_wrapper,
-    mock_data,
-    error_class,
-):
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        side_effect=[make_openai_error(error_class) for _ in range(3)]
-    )
-
-    result = await async_openai_wrapper.generate_topic_cluster_names(
-        ["test prompt"],
-        [mock_data["old_names"]],
-    )
-
-    assert result == [mock_data["old_names"]]
-
-
-@pytest.mark.asyncio
-@pytest.mark.filterwarnings("ignore:Failed to generate")
-async def test_async_openai_retries_per_item_not_whole_batch(
-    async_openai_wrapper, mock_data
-):
-    good_response = MockAsyncResponse.create_chat_response(
-        mock_data["valid_topic_name"]
-    )
-    error_class = OPENAI_RETRYABLE[0]
-    call_counts = {"prompt1": 0, "prompt2": 0}
-
-    async def mock_create(*args, **kwargs):
-        prompt_text = kwargs["messages"][0]["content"]
-        call_counts[prompt_text] += 1
-        if prompt_text == "prompt1":
-            return good_response
-        elif prompt_text == "prompt2":
-            raise make_openai_error(error_class)
-        raise AssertionError(f"Unexpected prompt: {prompt_text}")
-
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        side_effect=mock_create
-    )
-    result = await async_openai_wrapper.generate_topic_names(["prompt1", "prompt2"])
-
-    assert len(result) == 2
-    validate_topic_name(result[0])
-    assert result[1] == ""
-    assert call_counts["prompt1"] == 1
-    assert call_counts["prompt2"] == 3
-
-
-@pytest.mark.asyncio
-async def test_async_openai_generate_topic_names_retry_exhausted_warns(
-    async_openai_wrapper,
-):
-    error_class = OPENAI_RETRYABLE[0]
-
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        side_effect=[make_openai_error(error_class) for _ in range(3)]
-    )
-
-    with pytest.warns(UserWarning, match="Failed to generate topic name"):
-        result = await async_openai_wrapper.generate_topic_names(["test prompt"])
-
-    assert result == [""]
-
-
-@pytest.mark.asyncio
-async def test_async_openai_generate_topic_cluster_names_retry_exhausted_warns(
-    async_openai_wrapper, mock_data
-):
-    error_class = OPENAI_RETRYABLE[0]
-
-    async_openai_wrapper.client.chat.completions.create = AsyncMock(
-        side_effect=[make_openai_error(error_class) for _ in range(3)]
-    )
-
-    with pytest.warns(UserWarning):
-        result = await async_openai_wrapper.generate_topic_cluster_names(
-            ["test prompt"],
-            [mock_data["old_names"]],
-        )
-
-    assert result == [mock_data["old_names"]]
 
 
 # AsyncAzureAI Tests
